parser: upgrade to bison-3.5
Dust off the parser setup, so we can generate with recent bison.  This
is a really big change, mostly outside of the parser to glue the bits
together again.  It mostly involves the introduction of a HTMLDriver
class, that replaces the MyParser class that extended the HTMLControl
class (basically the scanner class).