istr.h: avoid warning from clang on 'always true' expression
In some cases we use the macro with a char instead of int, which makes
the static analysis true that the condition is pointless.  Add a cast in
the macro to make the argument an int.  Not sure if that makes clang
fully happy though.

Issue: https://github.com/grobian/html2text/issues/35
Signed-off-by: Fabian Groffen <grobian@gentoo.org>