*: licence changes #39, whitespace/comment changes
A confusing situation where conflicting licences were used in several
files.  For this decision, we're referring to a conversation with GMRS
Software/original authors where they agree to relicence the code to GPL.
This conversion is included in the HISTORY file.  As such, all files are
stripped form previous licences and now licenced under the GPL-2
licence.

While at it, remove a lot of whitespace and comments that are redundant,
or kept in Git's logs.

Closes: https://github.com/grobian/html2text/issues/39
Signed-off-by: Fabian Groffen <grobian@gentoo.org>