*: introduce ANSI-escape code rendering
Using ANSI escapes is much more Cool(tm) than using backspace sequences.
Not only is it more efficient, it's also handled by most terminals these
days, and it works much better with UTF-8 sequences, etc.  On top of
that, one can do COLOURS, yay!  So in this commit a bunch of things are
combined on top of each other:

1. ANSI escape handling
2. colour handling, kind of messy, but sometimes works well (e.g. links)
3. table reflow handling to pick up the properties for colours and
   alignment

Signed-off-by: Fabian Groffen <grobian@gentoo.org>